+++ example/lib/main.dart
@@ -34,7 +34,7 @@ class MyHomePage extends StatelessWidget {
       body: const Center(
         child: CreditCardUi(
           cardHolderFullName: 'John Doe',
-          cardNumber: '4123456781234567',
+          cardNumber: '4344343434343',
           validFrom: '01/23',
           validThru: '01/28',
           topLeftColor: Colors.blue,